Optimizations (changed ints to shorts, etc.)

+++ Core/datadefs.h
@@ -3,7 +3,7 @@
 
 extern vector<vector<int>> segdefs;
 extern vector<transdef> transdefs;
-extern unordered_map<string, int> nodenames;
+extern unordered_map<string, uint16_t> nodenames;
 extern vector<vector<vector<int>>> palette_nodes;
 extern vector<vector<vector<int>>> sprite_nodes;
